renamed About to Home and Home to Map
